requirements.txt
streamlit
python-dotenv
requests
supabase
postgrest
pandas
